# Conflicts:
#	SRC/ESINF/src/main/java/Main.java
#	SRC/ESINF/src/main/java/Structure/GraphAlgorithms.java
#	SRC/ESINF/src/main/java/US03/MinimumRoute.java
#	SRC/ESINF/src/main/java/US04/MinSpanningTreeResult.java
#	SRC/ESINF/src/test/java/Structure/GraphAlgorithmsTest.java

+++ SRC/ESINF/src/main/java/FileReader/ReaderFiles.java
@@ -24,7 +24,7 @@ public class ReaderFiles {
 
         while ((currentLine = reader.readLine()) != null){
             data = currentLine.split(",");
-            network.addRoute(new Hub(data[0]), new Hub(data[1]), Double.parseDouble(data[2]));
+            network.addRoute(new Hub(data[0]), new Hub(data[1]), Integer.parseInt(data[2]));
         }
     }
     /**
